+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - General
   - Consider `.lsp/config.edn` as part of project code, removing false positives of unused-public-var linter.
   - Consider full qualified symbols in edn files when checking for var references.
+  - Improving clojure-lsp linter capabilities, migrating `unused-public-var` and `different-aliases` linters to be built-in linters.
+    - Migrate from clj-kondo `custom-lint-fn` but considering kondo settings to avoid breaking changes.
 
 - Editor
   - Add support for LSP feature `textDocument/selectionRange`. #1961